Split collider hierarchies into plugin, add `AncestorCollider` marker
